finish frontend

+++ app/src/components/Item.jsx
@@ -5,7 +5,7 @@ export default function Item({ name, logo, amount, setAmount, id }) {
   function setA(v) {
     console.log(amount);
     let _amount = amount;
-    _amount[id] = parseInt(v.target.value);
+    _amount[id] = parseFloat(v.target.value);
     setAmount(_amount);
   }
